Javascript new fucntion and loop the interaction
I made a new function that encompases the whole code so that when I call it at the end of the function I can look and rerun the code . To make it more appealing I added a confirmation box so that when I select playAgain it will call the function playGame and restart everything.
One problem I have is now the dialog says what the computer chose but then changes it on the next dialog.