Javascript code deletion, move, and additions
I changed the location of console code and expanded on the human and computer choice from an if statement to an if else statemetn to make it where if the human didnt win then the computer won but only if both of their scores are different and it would then increment those scores. Then I removed the const human selection and computer selection because they over complicate the process by initializing new choices that will then lead to different results.